+++ resources/js/Pages/Admin/Teams/Team.vue
@@ -5,9 +5,9 @@ import AdminTopNavigation from "@/Components/Admin/AdminTopNavigation.vue";
 import {onMounted, ref} from "vue";
 import PaginatorComponent from "@/Components/Admin/PaginatorComponent.vue";
 import Swal from "sweetalert2";
-import AdminTeamMerchantTeamsComponent from "@/Components/Admin/AdminTeamMerchantTeamsComponent.vue";
+import AdminTeamMerchantTeamsComponent from "@/Components/Admin/TeamMerchantTeams/AdminTeamMerchantTeamsComponent.vue";
 import AdminTeamDetailsComponent from "@/Components/Admin/AdminTeamDetailsComponent.vue";
-import AdminTeamServiceTeamsComponent from "@/Components/Admin/AdminTeamServiceTeamsComponent.vue";
+import AdminTeamServiceTeamsComponent from "@/Components/Admin/TeamServiceTeams/AdminTeamServiceTeamsComponent.vue";
 import AdminUserSelectComponent from "@/Components/Admin/AdminUserSelectComponent.vue";
 
 const $props = defineProps({
